consistent formatting
I dunno if you want to include the ? in the `data?` key/field, but it's 
easier on my eye to have the three consistently formatted - mostly just 
cuz they are written about out of order.  Also - prolly my 
misunderstanding, are the three keys in the export type 
TransationInstructionCtorFields object supposed to be separated by semi 
colons or commas?