Over-used variable, "mint"
I updated the variable name in the "CreateNewMint()" function 
declaration to "tokenMint".

This way the calls to "mint" within the main() function stay 
minty-fresh.  The scope difference is acceptable, i.e. compiles and runs 
just fine, but in terms of conceptual clarity, I dislike re-using "mint" 
in two different scopes and this renaming also clarifies the use of 
"Token Mint"