fix: patrick's edits (#101)
* fix: remove list item end punc

* sugg: separate

2nd sentence --> new list item

* sugg: awk

"known data" reads kinda... dirty?

https://youtu.be/CtwJvgPJ9xw

* sugg: awk

The general statement re: plural "programs" makes the possessive 
ambiguous

* sugg: wot --> wot?

* sugg: awk

* sugg: awk

* sugg: awk

* sugg: awk

* sugg: awk

* sugg: clarify

* sugg: follow suit

I don't know that 255 initially needs to be code text, but if we're 
going there, seems like the following count should be same?

* sugg: awk punc fix opt1

etc.). --> et cetera).

Could also do

etc.). --> etc.) Once

* sugg: awk

I don't love this update, but it feels better?

* sugg: awk fix

* sugg: runon fix

Also, pretty sure we want to capitalize the E in Ed25519

* fix: ed --> Ed25519

* sugg: clarity

somewhat overstating, but given the binary "in/valid PDA" vs "on/off 
curve" I think it is worth doing.

* sugg: brevity

* sugg: brvty

* sugg: emph

* sugg: emph

strong caveat

* sugg: awk

honestly, this pg is not great - starts off with "One point" but has a 
lot of awk conjunctions

* sugg: flow

Also, I h8 the end line "tends to be the way to go" - why?  what about 
it is an advantage to non-PDA accounts?

* sugg: for --> of

* sugg: awk

* fix: repetitive

* sugg: localizing

traditional vs web3

* sugg: awk pg fix

* fix: sp
unecessary --> unnecessary

* sugg: awk runon

* sugg: awk

I don't love this as an ending line.  needs something to localize the 
examples for demo purpose - a la, "we don't recommend it.  We're just 
using the example for now to illustrate the mapping possibilities.

* sugg: y not --> y not?

* sugg: flow

* fix: let's --> lets

* sugg: non standard code breaks

to emphasize the array of seeds?

* sugg: h3

* sugg: awk fix

* fix: ordered h3

5 --> 7

* sugg: awk

* fix: awk grmr

* sugg: ol --> ul

unless there is a specific order, I think bullet points suffice

* sugg: awk

Maybe I am just a little loopy

* fix: sp
commments --> comments

* sugg: code text

* sugg: awk grmr fix

Co-authored-by: James <jamesrp13@gmail.com>